requirements.txt
PyYAML==5.1
Flask_Cors
Flask
python-dotenv
Flask-SQLAlchemy
PyJWT
waitress